refactor time range selector
